requirements.txt
gymnasium>=0.29
numpy>=1.24
scipy>=1.10
pyyaml>=6.0
stable-baselines3>=2.3.0
tensorboard>=2.15
pytest>=7.0
tqdm>=4.66
